Skip expired backfills instead of returning error
If a single backfill has expired it would cause the whole FetchMatches
call to error out. Now the backfill will be skipped instead similar to
the NotFound case. The status code returned for attempting to
acknowledge or update an expired backfill is changed from Unavailable
since the operation is not retryable.